Load plugins from config file
I believe it woujld be easier if there was a key in the main ini file with a list of enabled plugins.

With this PR, you can add a key `plugins` to the section `Program`:

```ini
[Program]
loglevel: debug
logfile: BS440.log
plugins: BS440csv, BS440mail
```

In my opinion this is easier to manage especially when you are testing out new/updated plugins.