Bs410 (#61)
* Dynamic handles

A BS410 uses different handles for the interesting characteristics. Look
up the correct handles to be used when evaluating the indications

* Increase connection timeout

Connection timeout needs to be increase from 5 to 8 seconds on a BS410.
No problem to do this for BS44x too.

* Add Device model to config file and use it to set addressType

BS410 uses public address type instead of random addressType like the
other BS series scales

* Also make this write to handle characteristic dynamic

Found another spot where a write to a fixed handle was done instead of a
looked-up handle

* BS410 uses different timestandard

It seems the BS410 uses seconds from 1/1/2010 instead of the secons
since epoch (1/1/1970) which is considered the Unix timestamp. The
specific standard for the BS410 is written to the scale. For now the
plugin builder needs to accomodate for this when decoding the data

* Fixed typo's

Various typos

* Swapped body and weight characteristics

Swapped the characteristics for body and weight data by mistake.

* Avoid looking up Handles on every indication

Store found handles once

* Fix valid data check

No need to compare again handle and expected handle. If they differ the
decode would not have been called anyway.

* Typo in setting command

again searched handle where handle was already available

* With BS410 Send correct time to plugins

When using BS410 not only send the correct time to the device but also
correct (i.e. convert to unix timestamp) the received timestamp as sent
to the plugins. Also got rid of decode.py as a separate file. Serves no
purpose to separate it from the rest of the code.